// src/app/modulos/portal/portal.enum.ts
export enum PaginasClave {
  Asociado = 'asociado',
  AsociadoReactivaciones = 'reactivaciones-detalle',
  BandejaSeguimientoEvaluacion = 'bandejaseguimientoevaluacion',
  Bandejacelula = 'bandejacelula',
  Bandejareactivaciones = 'bandejareactivaciones',
  EditarAsociado = 'editar-asociado',
  Miespacio = 'miespacio',
  Simulador = 'simulador'
}

export enum Perfiles {
  Administrador = 'ROLE_ADMINISTRADOR',
  AsistenteLegal = 'ROLE_ASISTENTE_LEGAL',
  AsistenteRecupero = 'ROLE_ASISTENTE_RECUPEROS',
  FuncionarioAutomotriz = 'ROLE_FUNCIONARIO_AUTOMOTRIZ',
  FuncionarioCredito = 'ROLE_FUNCIONARIO_CREDITO',
  FuncionarioServicioVentas = 'ROLE_FUNCIONARIO_SERVICIO_VENTAS',
  GerenciaOperaciones = 'ROLE_GERENCIA_OPERACIONES',
  JefeRecupero = 'ROLE_JEFE_RECUPEROS',
  JefeNegocios = 'ROLE_JEFE_NEGOCIOS'
}

export enum DocumentosBotones {
  Anular = 'Anular',
  Aprobado = 'Aprobado',
  EnvioComite = 'Envío a comité',
  Observado = 'Observado',
  Siguiente = 'Siguiente'
}

export const BANDEJA_TRABAJO = 'Mi bandeja de trabajo';